data element simplifications
